requirements.txt
azure-ai-projects==1.0.0b12
azure-identity==1.23.1
azure-ai-agents==1.1.0b4
python-dotenv==1.1.1
requests==2.32.4
beautifulsoup4
mcp>=1.0.0
azure-storage-blob>=12.15.0
python-docx>=1.1.0
reportlab>=4.0.0
markdown>=3.5.0